Updated About Me content. Renamed portfolio to projects
