+++ src/components/main/about-me/index.js
@@ -5,22 +5,22 @@ export default function AboutMe({ uuidv4 }) {
     <article className='container'>
       
       <section>
-        <h2 className="article-header">Hi, I'm Erik.</h2>
+        <h2 className="article-header">Hi, I'm Erik</h2>
       </section>
       
       {/* <section> */}
       <section className="article-content-container">
 
-        <div className="article-content">
+        {/* <div className="article-content"> */}
           
-          <p className="article-content-title">
+          {/* <p className="article-content-title">
             Thanks for checking out my website.
-          </p>
+          </p> 
           <p className="article-content-indent">
-            Magna magna voluptate excepteur labore aliqua.
-          </p>
+            I've built this page as a portfolio, 
+          </p> */}
 
-          <p className="article-content-title">
+          {/* <p className="article-content-title">
             IT Engineer, MSP Director, Project Manager, and Software Developer. 
           </p>
           <p className="article-content-indent">
@@ -35,16 +35,19 @@ export default function AboutMe({ uuidv4 }) {
           <p className="article-content-indent">
             Time goes by fast and I want to make memories with the people I love.
           </p>
-        </div>
+        </div> */}
 
         <div className="article-content">
           <p className="article-content-title">
-            When it comes to hobbies, software development is my focus lately.
+            Welcome to my professional portfolio.
           </p>
           <p className="article-content-indent">
-            To learn more about my projects, check out my Portfolio.
-          </p>          
-          <p className="signature">- Erik Plachta on March 20th, 2022</p>
+            Check out the <b>Projects</b> tab to learn about what projects I"m working
+            on. Check out my <b>Resume</b> to learn about my professional experience.
+            If you want to get in touch, head on over to the <b>Contact</b> page or
+            reach out to me on my social links below.
+          </p>
+          <p className="signature">- Erik Plachta - March 2022</p>
         </div>
       </section>
     </article>
